project/requirements.txt
amqp==2.1.4
Babel==2.3.4
billiard==3.5.0.2
celery==4.0.2
Django==1.10.4
django-braces==1.10.0
django-cors-headers==1.3.1
django-filter==1.0.1
django-oauth-toolkit==0.11.0
django-rest-swagger==0.3.9
djangorestframework==3.5.3
flower==0.9.1
gunicorn==19.6.0
kombu==4.0.2
Markdown==2.6.7
oauthlib==1.1.2
psycopg2==2.6.2
pytz==2016.10
PyYAML==3.12
redis==2.10.5
six==1.10.0
tornado==4.2
vine==1.1.3
